#703 More refactoring
